change changeUserLevel to changeLevel

+++ userManager.class.php
@@ -205,7 +205,7 @@
 		}
 		
 		//Checks if the given user is an admin
-		private static function isAdmin($userID){
+		private function isAdmin($userID){
 			$parameters=Array();
 			$parameters[":userID"]=$userID;
 			//TODO: Following two lines probably have a shortcut
@@ -218,40 +218,50 @@
 			if(!$this->getLoginState())
 				return false;		
 			$current=Array();
-			$current[":userID"]=getSession()["ID"];
-			if(isAdmin(getSession()["ID"])){
+			$current[":userID"]=$this->getSession()["ID"];
+			if($this->isAdmin($this->getSession()["ID"])){
 				if(!array_values($this->DB->getRow("SELECT Deleteable FROM Admins WHERE UserID = :userID", $current))[0])
 					return true;
 				else
-					return !isAdmin($userID);
+					return !$this->isAdmin($userID);
 			}else
 				return false;
 		}
-
-		//Promotes or demotes a user according to the given level(0=user, 1=admin, 2=god)
-		public function changeUserLevel($userID, $level){
-			$current=Array();
-			$current[":userID"]=getSession()["ID"];
+		
+		private function isDeleteable($userID){
 			$parameters=Array();
 			$parameters[":userID"]=$userID;
-			$isThisDeleteable=array_values($this->DB->getRow("SELECT Deleteable FROM Admins WHERE UserID = :userID"),$current)[0];
-			if($this->isSuperrior($userID)){
+			return array_values($this->DB->getRow("SELECT Deleteable FROM Admins WHERE UserID = :userID",$parameters))[0];
+		}
+
+		//Promotes or demotes a user according to the given level(0=user, 1=admin, 2=god)
+		public function changeLevel($userID, $level){
+			$current=Array();
+			$current[":userID"]=$this->getSession()["ID"];
+			print_r($current);
+			$parameters=Array();
+			$parameters[":userID"]=$userID;
+			if($this->isSuperior($userID)){
 			//TODO: add confirmation messages
 				switch($level){
 					case 0:
-						if(isAdmin($userID)){
-							$this->DB->query("DELETE FROM " . ADMIN_TABLE . "WHERE UserID = :userID", $parameters);
+						if($this->isAdmin($userID)&&$this->isDeleteable($userID)){
+							$this->DB->query("DELETE FROM " . ADMIN_TABLE . " WHERE UserID = :userID", $parameters);
+							return true;
+						}else{
+							return false;
 						}
-						return true;
 					case 1:
-						if(!isAdmin($userID)){
+						if(!$this->isAdmin($userID)){
 							$this->DB->query("INSERT INTO " . ADMIN_TABLE . "(UserID, Deleteable) VALUES (:userID, 1)", $parameters);
+							return true;
+						}else{
+							return false;
 						}
-						return true;
 					case 2:
-						if(!$isThisDeleteable){
-							if(isAdmin($userID)){
-								$this->DB->query("UPDATE " . ADMIN_TABLE . "SET Deleteable = 0 WHERE UserID = :userID", $parameters);
+						if(!$this->isDeleteable($current[":userID"])){
+							if($this->isAdmin($userID)){
+								$this->DB->query("UPDATE " . ADMIN_TABLE . " SET Deleteable = 0 WHERE UserID = :userID", $parameters);
 							}else{
 								$this->DB->query("INSERT INTO " . ADMIN_TABLE . "(UserID, Deleteable) VALUES (:userID, 0)", $parameters);
 							}
